Removed Json.NET dependency.
