--- src/Chinchilla.Api/LowercaseContractResolver.cs
@@ -1,12 +1,0 @@
-using Newtonsoft.Json.Serialization;
-
-namespace Chinchilla.Api
-{
-    public class LowercaseContractResolver : DefaultContractResolver
-    {
-        protected override string ResolvePropertyName(string propertyName)
-        {
-            return propertyName.ToLower();
-        }
-    }
-}
