+++ src/Chinchilla.Api/packages.config
@@ -1,6 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Newtonsoft.Json" version="5.0.8" targetFramework="net40" />
   <package id="RestSharp" version="104.3.3" targetFramework="net45" />
   <package id="SpeakEasy" version="0.2.0.0" targetFramework="net45" />
 </packages>
